feat: restructure infra components to streamline (#28)
Prepares for production deployments of the services out of experimental.

A connector comes with
```
service
├── deploy
│   ├── helm-charts
│   │   └── service
│   └── terraform-azure
│       ├── service-scope-module1
│       └── service-scope-module2
```

## helm-charts
is plural, you never know

## modules
differ per service, thus always are `service-name-XXXX` where XXXX is
first a scope then a function.

## entra module
is not yet deployed and won't be until 5min before deployment where
permissions and everything else are crisp and fresh and documented.