feat(sharepoint-connector): UN-13757 Sharepoint  connector existing business logic (#33)
This PR brings all the logic from the initial monorepo migration to the
new connectors repository.
We are not deploying anything. This is to merge the business logic in so
I can make smaller PRs with the new features we'll be adding.

There have been a lot of changes done to it since it's first days:
- no longer using queues, we're scanning a sharepoint site, process it
and then continue to the next site.
- now we're using ms graph sdk with all the middlewares it requires -
like it's implemented in outlook mcp
- removed all custom metrics previously added for processing and
scanning.
- added open telemetry. At the moment it is not currently utilized, this
will change in ticket: https://unique-ch.atlassian.net/browse/UN-13753
- removed healthcheck endpoint, project now usese @unique-ag/probe as
outlook-mcp
- project is using biome and vitest 
- path based ingestion will be implemented in
https://unique-ch.atlassian.net/browse/UN-13777
- graphql types will be implemented in
https://unique-ch.atlassian.net/browse/UN-13756
- refactored a lot of the code to improve readability of the services
- added ms graph api rate limiter of 10k requests per 10 seconds. 
- added min file size check, currently checking for 0 bytes file sizes. 
- fixed: correctly setting the file title when ingesting


Service tested locally on our QA environment and sharepoint instance.

---------

Co-authored-by: Lorand Sandor <lorand@Lorands-MacBook-Pro.local>