fix(sharepoint-connector): do not request Graph API token multiple times
Because the token is not populated before request comes back, it caused
us to request new token several times until the cache got populated. We
switched to cache promise instead of direct value, so that after the
first time we request the token, we have a promise to use for the
consecutive requests.