fix(sharepoint-connector): do not request Graph API token multiple times (#79)
Because the token is not populated before request comes back, it caused
us to request new token several times until the cache got populated. We
switched to cache promise instead of direct value, so that after the
first time we request the token, we have a promise to use for the
consecutive requests.

As part of this PR we also cleaned up and fixed up some Github Workflows.
There was concurrency grouping problem and templates should not longer
indicate being MCP-specific.