feat(sharepoint-connector): support for Unique API auth modes
I've split Unique config into discriminated union, because whole Zitadel
configuration is redundant for cluster_local auth mode, while headers
are ignored if we use the external auth mode.

I renamed the Zitadel env variables to keep them in line with agreed
ZITADEL_SERVICE_EXTRA_HEADERS. It's a small price for some consistency.

Because existing client for Unique APIs required to pass the token, I've
decided to move the file ingestion endpoints to a new service. It is
based on the same principles as the groups and users services, but
implements HTTP file-diff as well.

During testing it came out that we need to set also roles header to be
able to perform some operations via GraphQL so I added the header with
the permissions to the documentation and examples.

I removed the HttpModule because it was a bit of a weird one, because
each client was used for one specific case, so I moved the clients to
services that want to use them. We can go back to the previous formula
when we find a case when one client is reused across multiple other
services.