feat(sharepoint-connector): allow auto-creating key and certificate (#104)
## Risk Triage
This will put the key into the terraform state. 

### Internal mitigations

The key appears in the state.

- Use `ephemeral` resource - does not work as the `certificate` does not
allow `wo` `ephemeral` input
- Make certificate creation optional / opt-in - done
- Notify users in variables - done
- Offer rotation - done

Further risk considerations: In most companies (incl. Unique) the attack
vector to reach the terraform state is more complex than reaching the
Key Vault itself (where the Key ends up). This is due to the fact that
Support teams support clients and potentially debug an issue. Support
teams never need to access the terraform state itself for their daily
work.

### External Mitigations
Users can use `Named Locations` and `Conditional Access Policies` to
lock the use of the certificate to a source IP-Range or Location.

## Note on rotation
**Will there be downtime?** Not if its pre-expire because the pods will
still know the old key until rebooted.
Real, as code rotation is something that can be tackled in future
releases.

Refs: UN-14890