Clean paths creation and handling
Previous implementation did lot's of pretty weird and seemingly
unnecessary things without any explanation as to what it does. New
implementation is order of magnitude simpler.

New implementation uses path fir leading slash to be in line with how
absolute paths are built in file system.

We also separated the interfaces of Scope and ScopeWithPath to be clear
of which variant is passed where and improve handling within the type
system. Now we know what we get from the API and what we enrich with
path information.