feat(sharepoint-connector): experimental re-write of MS Graph API client
We are having troubles with requests for content when there are multiple
smaller files involved. We suspect the problem is bound within the
library that seems to be abandoned without any releases in last 2 years.

This implementation was entirely written by AI agent just to confirm our
suspicion, that this library is the issue with requests on QA, proper
review and fixes will come next.