refactor(sharepoint-connector): update README and values.yaml for env… (#152)
…ironment variable handling

Revised documentation in README.md to clarify the usage of environment
variables from secrets, specifically for the
UNIQUE_ZITADEL_CLIENT_SECRET when authMode is set to 'external'. Updated
values.yaml to reflect this change, removing the hardcoded secret and
providing clearer instructions for users. Additionally, added context to
the certificate-values.yaml regarding the purpose of the secret for
SharePoint authentication.