feat(sharepoint-connector): prevent accidental deletion of all files (#153)
If the file diff indicated we should delete all files, it most probably
means that we have some kind of bug and we should not proceed with the
sync to avoid costly re-ingestions. In case user actually wants to
delete all files, they should follow instructions in the log depending on
their case.

Refs:
- https://unique-ch.atlassian.net/browse/UN-15140